@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the ONGR package.
  *
@@ -18,136 +20,108 @@ use ONGR\FilterManagerBundle\SerializableInterface;
  */
 class ChoiceAwareViewData implements SerializableInterface
 {
-    /**
-     * @var bool
-     */
-    private $active = false;
+
+    private bool $active = false;
+
+
+    private bool $default = false;
 
     /**
-     * @var bool
+     * Holds set or unset parameters depending on state.
      */
-    private $default = false;
+    private array $urlParameters = [];
 
     /**
-     * @var array Holds set or unset parameters depending on state.
      */
-    private $urlParameters = [];
+    private string $label;
 
     /**
-     * @var string
+     * Sorting any arrays: "min", "max", for only numeric arrays: "avg", "sum".
      */
-    private $label;
+    private ?string $mode = null;
 
     /**
-     * @var string Sorting any arrays: "min", "max", for only numeric arrays: "avg", "sum".
+     * Represents document count for option.
      */
-    private $mode = null;
+    private int $count = 0;
 
-    /**
-     * @var int Represents document count for option.
-     */
-    private $count = 0;
 
-    /**
-     * @return bool
-     */
-    public function isActive()
+    public function isActive(): bool
     {
         return $this->active;
     }
 
-    /**
-     * @param bool $active
-     */
-    public function setActive($active)
+
+    public function setActive(bool $active): void
     {
         $this->active = $active;
     }
 
-    /**
-     * @return bool
-     */
-    public function isDefault()
+
+    public function isDefault(): bool
     {
         return $this->default;
     }
 
-    /**
-     * @param bool $default
-     */
-    public function setDefault($default)
+
+    public function setDefault(bool $default): void
     {
         $this->default = $default;
     }
 
-    /**
-     * @return int
-     */
-    public function getCount()
+
+    public function getCount(): int
     {
         return $this->count;
     }
 
-    /**
-     * @param int $count
-     */
-    public function setCount($count)
+
+    public function setCount(int $count)
     {
         $this->count = $count;
     }
 
-    /**
-     * @return string
-     */
-    public function getLabel()
+
+    public function getLabel(): string
     {
         return $this->label;
     }
 
-    /**
-     * @param string $label
-     */
-    public function setLabel($label)
+
+    public function setLabel(string $label): void
     {
         $this->label = $label;
     }
 
-    /**
-     * @return array
-     */
-    public function getUrlParameters()
+
+    public function getUrlParameters(): array
     {
         return $this->urlParameters;
     }
 
-    /**
-     * @param array $urlParameters
-     */
-    public function setUrlParameters($urlParameters)
+
+    public function setUrlParameters(array $urlParameters): void
     {
         $this->urlParameters = $urlParameters;
     }
 
-    /**
-     * @return string
-     */
-    public function getMode()
+
+    public function getMode(): ?string
     {
         return $this->mode;
     }
 
-    /**
-     * @param string $mode
-     */
-    public function setMode($mode)
+
+    public function setMode(string $mode = null): void
     {
         $this->mode = $mode;
     }
 
+
     /**
      * {@inheritdoc}
      */
-    public function getSerializableData()
+    public function getSerializableData(): array
     {
         return [
             'active' => $this->isActive(),
