@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the ONGR package.
  *
@@ -23,7 +25,7 @@ class DateRange extends Range
     /**
      * {@inheritdoc}
      */
-    public function getViewData(DocumentIterator $result, ViewData $data)
+    public function getViewData(DocumentIterator $result, ViewData $data): ViewData
     {
         $name = $data->getState()->getName();
 
